+++ demo/src/app/app.routes.ts
@@ -34,7 +34,7 @@ import { DemoPropertyOperationSectionComponent } from './components/+property-op
 import { DemoUploaderSectionComponent } from './components/+uploader/uploader-section.component';
 import { DemoDateRangeSectionComponent } from './components/+date-range/date-range-section.component';
 import { DemoKeySelectSectionComponent } from './components/+key-select/key-select-section.component';
-import { DemoListComponent } from './components/+list/list-section.component';
+import { DemoListSectionComponent } from './components/+list/list-section.component';
 import { DemoDatepickerNextSectionComponent } from './components/+datepicker-next/datepicker-next-section.component';
 import { DemoTreeSelectSectionComponent } from './components/+tree-select/tree-select-section.component';
 import { DemoStrengthComponent } from './components/+strength/strength-section.component';
@@ -264,7 +264,7 @@ const appRoutes = [
     },
     {
         path: 'components/list',
-        component: DemoListComponent
+        component: DemoListSectionComponent
     },
     {
         path: 'components/tree-select',