+++ src/datepicker/datepicker-container.component.ts
@@ -1,7 +1,7 @@
-import { Component, OnInit, ViewChild, ElementRef, Renderer2, ViewContainerRef, ComponentRef, HostBinding } from '@angular/core';
+import { Component, ElementRef, OnInit, Renderer2, ViewChild, ViewContainerRef } from '@angular/core';
 import { BsDatepickerConfig } from 'ngx-bootstrap/datepicker';
-import { DatepickerValueEntry } from './i.datepicker';
 import { skip } from 'rxjs/operators';
+import { DatepickerValueEntry } from './i.datepicker';
 
 @Component({
     selector: 'thy-datepicker-container',
@@ -21,7 +21,12 @@ export class ThyDatepickerContainerComponent implements OnInit {
     @ViewChild('dpContainer')
     private _dpContainerRef: any;
 
-    constructor(private _config: BsDatepickerConfig, _elementRef: ElementRef, _renderer: Renderer2, _viewContainerRef: ViewContainerRef) {
+    constructor(
+        private _config: BsDatepickerConfig,
+        _elementRef: ElementRef,
+        _renderer: Renderer2,
+        _viewContainerRef: ViewContainerRef
+    ) {
         this._config.containerClass = 'theme-ngx';
         this._config.showWeekNumbers = false;
     }
@@ -82,7 +87,9 @@ export class ThyDatepickerContainerComponent implements OnInit {
     }
 
     private _initDataValue() {
-        this.value = this.initialState.value.date ? new Date(this.initialState.value.date.getTime()) : new Date();
+        this.value = this.initialState.value.date
+            ? new Date(this.initialState.value.date.getTime())
+            : this.initialState.defaultDate || new Date();
         this._dpContainerRef._effects.init(this._dpContainerRef._store);
         this._dpContainerRef._effects.setValue(this.value);
         this._dpContainerRef._effects.setMinDate(this.initialState.minDate);